Changed aggregate declaration to constructor, removed printFullArray
Organized imports, made aggregate setable in constructor, removed
superfluous printFullArray variable, and made the aggregate variable a
solid part of the parameters.
Have begun work on a function to load in parameters from a string in a
file, such that a single file with a lot of parameters for simulations
can be run.