Reworked a lot of the simulator
Changed to use scanners instead of buffered reader. Made the parameters
have a base element for the simulator to use when generating new
configuration. Changed so that division by nAtoms only happens once.
Changed all nXnYnZ to be number of unit cells, instead of dimensions of
array. Corrected the calculation of variance, it was wrong before.
Also corrected mistake where a comma wasn't put after Gzvar. Also
changed so that file handling is better.
And finally cleaned up to make it easier to look at.