Has moved the baseStates to a separate Enum
The baseStates are now in a separate Enum, and have been checked that
they behave correctly, by comparing with the earlier Crystal baseStates.
The baseStates has also been iterated from having an array of vectors,
to just a list of plus or minuses, and an integer determining which
coordinate to use. The setCoord of MyVector has been implemented. The
simulator now also only calculates the difference in basis state
projection, which makes simulations run much faster, does not depend on
system size. Probably also some other changes. Will checks to see if
everything looks okay now.