Changes to the directory system and
Changed name of H to B. Now also outputs the spin of the elements. Now
the simulation does not refer to the root directory, which is required
for the script to run on the cluster. Also stuff is more neatly
organized. Main script also outputs command and a config file can be
loaded with loadConfig.