Higher order moments
Has implemented that the third moment Expectation[E^3] and fourth
moments Expectation[F^4], Expectation[C^4], Expectation[A^4],
Expectation[G^4], are calculated. The third order moments of energy are
calculated for determining the derivative of the heat capacity, and the
fourth order moments of the basis states are calculated to obtain the
Binder cumulant.