Expected running time and simulator update
Has implemented a function for calculating the expected running tim eof
the program, and the simulator now has a alternative iteration
technique, where neighbouring atoms are switched. The old iteration
function for the simulator has also been updated; it should be checked
both that the new itaration technique and the update to the iterator
works as intended.